Add http_s package with handleTunneling function
This commit adds the http_s package along with the handleTunneling function. The handleTunneling function is responsible for establishing a tunnel between the client and the target server. It first dials the target server using the net.DialTimeout function, and if successful, it hijacks the response writer using the http.Hijacker interface. It then transfers data between the client and the target server using the transfer function.

Signed-off-by: SammyOina <sammyoina@gmail.com>