Fix bug in http.Handler()
The commit fixes a bug in the http.Handler() function where the request body was not being closed properly. This caused a resource leak. The bug is fixed by adding a call to r.Body.Close() to close the request body. Additionally, the request body is replaced with a new io.NopCloser(bytes.NewBuffer(payload)) to allow subsequent reads of the body.

Signed-off-by: SammyOina <sammyoina@gmail.com>