Fix authorization handling in websockets handler
This commit fixes the authorization handling in the websockets handler.
Previously, the code was assuming that the authorization token would always be provided as a query parameter. However, this is not always the case, as the token can also be provided in the `Authorization` header.

To fix this issue, the code now checks both the query parameter and the header for the authorization token. If the token is not found in either of these locations, an `ErrAuthorizationNotSet` error is returned.

This change ensures that the websockets handler correctly handles authorization and provides a more robust and secure implementation.

Signed-off-by: SammyOina <sammyoina@gmail.com>