Fix bug in CoAP proxy handler
The CoAP proxy handler had a bug where the path was not correctly extracted from the request options. This caused errors when trying to send the request to the target connection. This commit fixes the bug by correctly extracting the path and using it to make the request to the target connection.

This commit also adds a new method, `observeUpstream`, to handle CoAP observe requests. This method extracts the path from the request options and sets up a callback function to handle incoming observe notifications from the target connection.

The bug fix and new method improve the reliability and functionality of the CoAP proxy handler.

Signed-off-by: SammyOina <sammyoina@gmail.com>