Fix issue with request body not being reset after reading it in the HTTP handler
The HTTP handler in `pkg/http/http.go` was not resetting the request body after reading it, which caused subsequent calls to fail. This commit fixes the issue by setting the request body to a new `io.NopCloser` with the payload buffer. Now, the request body can be read multiple times without any issues.

Signed-off-by: SammyOina <sammyoina@gmail.com>