UV-267 -  WS proxy support (#44)
* add ws proxy adapter

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Refactor websocket package and update imports

The websocket package has been refactored to improve code organization and maintainability. The following changes have been made:

- The package import path has been updated to "github.com/mainflux/mproxy/pkg/websockets".
- Unused constants and variables related to the websocket configuration have been removed.

These changes improve the clarity and readability of the codebase.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix authorization handling in websockets handler

This commit fixes the authorization handling in the websockets handler.
Previously, the code was assuming that the authorization token would always be provided as a query parameter. However, this is not always the case, as the token can also be provided in the `Authorization` header.

To fix this issue, the code now checks both the query parameter and the header for the authorization token. If the token is not found in either of these locations, an `ErrAuthorizationNotSet` error is returned.

This change ensures that the websockets handler correctly handles authorization and provides a more robust and secure implementation.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix import error in websockets package

Import the missing package "golang.org/x/sync/errgroup" to fix the import error in websockets.go file.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* export handler

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Refactor configuration struct and update WebSocket proxy logging

The commit refactors the configuration struct in the main.go file
to improve readability and maintainability. The struct fields
(httpConfig, mqttConfig, wsMQTTConfig, wsConfig, and logLevel)
are now properly aligned and indented.

Additionally, the commit updates the logging message for starting
the encrypted WebSocket proxy on port cfg.wsMQTTConfig.wssPort.
The previous message incorrectly referenced cfg.mqttWssPort.

This commit follows the best practices for writing commit messages
and provides a clear and concise summary of the changes made in
the commit.

Signed-off-by: SammyOina <sammyoina@gmail.com>

* Fix incorrect variable assignment in Proxy.Handler()

The variable "target" in the Proxy.Handler() function was assigned
using an incorrect format string. This commit fixes the issue by
updating the format string to the correct value.

Summary:
- Fix incorrect variable assignment in Proxy.Handler()

Body:
- The variable "target" in the Proxy.Handler() function was assigned
  using an incorrect format string. This commit fixes the issue by
  updating the format string to the correct value.

Signed-off-by: SammyOina <sammyoina@gmail.com>

---------

Signed-off-by: SammyOina <sammyoina@gmail.com>